Created migrate module and moved stub migration tests
